+++ app/db/Streams.scala
@@ -17,4 +17,5 @@ trait Streams {
   def searchStreams(geocode: Option[String], stream_name: Option[String]): List[StreamModel]
   def getBinForStream(time: String, stream_id: Int): List[JsValue]
   def deleteStream(id: Int): Unit
+  def deleteStreams(start: Int, end: Int): Unit
 }